+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0
 - Back office : Initialiser une nouvelle société avec son administrateur
 - Ajout des liens vers twitter, linkedin, fb, eurekaci.com dans le footer
 - Permettre aux utilisateurs de recommander RDI-Manager
+- Notification de rappel des temps saisis également par SMS
 ### Modifié
 - Migration vers Symfony 5